Plando: Use PlandoAssignmentService instead of FormControl kludge.
