+++ app/src/app/pages/home/randomizer-page/goal-settings/goal-settings.component.html
@@ -12,7 +12,7 @@
       </div>
 
       <div class="setting-element">
-        <mat-slide-toggle formControlName='includePowerStars' plandoAssignable [plandoAssignedControls]="plandoAssignedControls"></mat-slide-toggle>
+        <mat-slide-toggle formControlName='includePowerStars' plandoAssignable></mat-slide-toggle>
         <app-tooltip-span spanText="Star Hunt"
                           tooltipText="Allows adding power stars into the item pools.
 
@@ -31,7 +31,7 @@
       </div>
 
       <div class="setting-element" [ngClass]="{ 'disabled-element': isSevenOrZeroStarSpirits}">
-        <mat-slide-toggle formControlName='requireSpecificSpirits' plandoAssignable [plandoAssignedControls]="plandoAssignedControls"></mat-slide-toggle>
+        <mat-slide-toggle formControlName='requireSpecificSpirits' plandoAssignable></mat-slide-toggle>
         <app-tooltip-span spanText="Require Specific Spirits"
                           tooltipText="Specific chapters must be completed to open the star way. You must visit shooting star summit to see which ones."></app-tooltip-span>
       </div>
@@ -47,7 +47,7 @@
         <mat-divider />
 
         <div class="setting-element">
-          <mat-slide-toggle formControlName='shuffleStarBeam' plandoAssignable [plandoAssignedControls]="plandoAssignedControls"></mat-slide-toggle>
+          <mat-slide-toggle formControlName='shuffleStarBeam' plandoAssignable></mat-slide-toggle>
           <app-tooltip-span spanText="Shuffle Star Beam"
                             tooltipText="Star Beam becomes an item shuffled in the item pool. Another random item will replace star beam in Star Haven."></app-tooltip-span>
         </div>
