+++ app/src/app/pages/home/randomizer-page/item-pool-settings/item-pool-settings.component.html
@@ -36,7 +36,7 @@
       </div>
 
       <div class="setting-element">
-        <mat-slide-toggle formControlName='itemPouches' plandoAssignable [plandoAssignedControls]="plandoAssignedControls"></mat-slide-toggle>
+        <mat-slide-toggle formControlName='itemPouches' plandoAssignable></mat-slide-toggle>
         <app-tooltip-span spanText="Add Item Pouches" tooltipText="Adds 5 item pouches into the item pool. Every pouch found increases Mario's inventory size by 2, so his total item capacity can reach a maximum of 20."></app-tooltip-span>
       </div>
 
@@ -51,7 +51,7 @@
       </div>
 
       <div class="setting-element">
-        <mat-slide-toggle formControlName='progressiveBadges' plandoAssignable [plandoAssignedControls]="plandoAssignedControls"></mat-slide-toggle>
+        <mat-slide-toggle formControlName='progressiveBadges' plandoAssignable></mat-slide-toggle>
         <app-tooltip-span spanText="Progressive Badges" tooltipText="Removes all jump charge, hammer charge, quake hammer badges, as well as Power Jump / Mega Jump and Power Smash / Mega Smash.
         Then, 3 progressive badges are added for each of those 5 badge families. Beta badges are inserted into the progressive sequence, even if beta badges are turned off."></app-tooltip-span>
       </div>
