+++ app/src/app/pages/home/randomizer-page/plandomizers/plandomizers.component.ts
@@ -1,16 +1,18 @@
-import { Component, Input, OnInit } from '@angular/core';
+import { Component, Input, OnDestroy, OnInit } from '@angular/core';
 import { FormControl } from "@angular/forms";
+import { Subscription } from "rxjs";
 import { SAVED_PLANDO_NAME_PREFIX, SAVED_PLANDO_NAMES_KEY } from "src/app/pages/plando-page/plando-save-load/plando-save-load.component";
+import { PlandoAssignmentService } from "src/app/services/plando-assignment.service";
 
 @Component({
   selector: 'app-plandomizers',
   templateUrl: './plandomizers.component.html',
   styleUrls: ['./plandomizers.component.scss']
 })
-export class PlandomizersComponent implements OnInit {
+export class PlandomizersComponent implements OnInit, OnDestroy {
   @Input() public plandomizerFormControl: FormControl;
-  @Input() plandoAssignedControls: FormControl<Set<string>>;
 
+  private _plandoAssignmentSubscription: Subscription;
   public readonly controlDisplayStrings: Map<string, string> = new Map([
     ["shuffleItems", "Shuffle Items"],
     ["includePanels", "Include Hidden Panels"],
@@ -41,14 +43,23 @@ export class PlandomizersComponent implements OnInit {
     ["addBetaItems", "Add Beta Items"],
     ["progressiveBadges", "Progressive Badges"],
   ]);
-
   public savedPlandoNames: Set<String>;
   public selectedPlandoName: string;
   public loadStatus: string;
+  public plandoAssignedControls: Set<string>
+
+  constructor(private _plandoAssignmentService: PlandoAssignmentService) { }
 
   public ngOnInit(): void {
     const storedNames = window.localStorage.getItem(SAVED_PLANDO_NAMES_KEY);
     this.savedPlandoNames = storedNames ? new Set(storedNames.split(',')) : new Set();
+    this._plandoAssignmentSubscription = this._plandoAssignmentService.assignedControls.asObservable().subscribe(assignedControls => this.plandoAssignedControls = assignedControls);
+  }
+
+  ngOnDestroy(): void {
+    if (this._plandoAssignmentSubscription) {
+      this._plandoAssignmentSubscription.unsubscribe();
+    }
   }
 
   public onSavedPlandoSelect($event: InputEvent) {
